Created projections for handling querying movies and rates.

+++ rating/src/main/java/pl/dsyou/movierating/rating/domain/RateFactory.java
@@ -1,12 +1,14 @@
 package pl.dsyou.movierating.rating.domain;
 
+import java.math.BigDecimal;
+
 public class RateFactory {
 
     public static Rate of(long movieId) {
         return new Rate(movieId);
     }
 
-    public static Rate of(long movieId, float rate) {
+    public static Rate of(long movieId, BigDecimal rate) {
         return new Rate(movieId, rate);
     }
 }
